Resize Device.label and rename the field to "Device identifier"
The internal service labels were limited to 4 characters, but
obviously this assumption is no longer valid so we set it to the
same as the hostname (200). Quite a few DB backends ignore it
anyway, but we need a DDL migration for those that don't.

Make DeviceType.tasks nullable (DB compatibility)