OpenShift development environment (#34)
Note: This pulls in mysqlclient, a python-mysqldb fork. It does
not interfere with existing production setups, but the build
dependencies for mysqlclient will have to be installed.

As discussed earlier, we don't want to depend on any distribution
packages except for the interpreter (portability).
This leaves us with only python-ldap.

Debian 8:

 - libmysqlclient-dev
 - python-dev

We'll need to fully migrate to mysqlclient for Python 3.