Make device locking mechanism more robust
Sometimes admins would run configmaster as root users which would lead
to lock files being inaccessible for subsequent runs of the normal
configmaster user. This is now avoided by explicitly making the
lockfiles readable and writable by anyone.

Also added a try-finally block to make sure that locks are always
released, as not doing so could lead to blocking and timeout if several
tasks would be run for one device with one task failing.

CON-41